cambios en la inversa a double
Se cambio los métodos de la inversa para recibir y enviar double

+++ src/main/java/com/matrices/Servicio_Matrices/controller/MatrizController.java
@@ -25,7 +25,7 @@ public class MatrizController {
 	
 	@CrossOrigin(origins = "*", methods= {RequestMethod.POST})
 	@PostMapping(path="/inversa")
-	public int[][] inversaMatriz(@RequestBody String matrices) {
+	public double[][] inversaMatriz(@RequestBody String matrices) {
 		return matrizService.calcularInversa(matrices);
 	}
